(Bug 22675) Fix DynamicPageList (3rd party) to work with mediawiki post r61913
All this does is make the extension work like it used to. However I find the aproach
this extension takes to included html to be rather scary. I'm a bit hesitent to commit this
as I think it might have security issues - however its no different from previous versions of this
extension. I fixed some of the obvious security issues in r68811 but I fear there might be more.