Add a new gallerysize parameter
### `gallerysize` Parameter Support

This adds support for a new `gallerysize` parameter, which allows users to control image sizing when using `mode=gallery`.

#### How it works:
- A single value (e.g., `gallerysize=100px`) sets both `widths` and `heights` to the same value.
- A comma-separated pair (e.g., `gallerysize=120px,90px`) sets `widths` and `heights` individually.

These values are passed directly to the `<gallery>` tag as `widths` and `heights` attributes.

#### Example:
```
<dpl>
mode=gallery
gallerysize=150px,100px
</dpl>
```

This would result in: `<gallery widths=150px heights=100px>...</gallery>`

If only a single value is provided, both width and height will use that value.


This also changes the behavior of `gallerymode` and if not using the parameter it won't set `mode` in the `<gallery>` tag at all. This is so that, by default, it can respect the configured default mode in `$wgGalleryOptions`.